Added more requirement checking functions
Added functions for checking Writing Intensive, ACCT, Upper Division, and Core courses if they are finished or not. Moved all these and gen ed checking function to RequirementsController.java
