--- h2/h2db.lock.db
@@ -1,6 +1,0 @@
-#FileLock
-#Thu Nov 14 13:03:58 EST 2019
-server=157.89.188.43\:64352
-hostName=157.89.188.43
-method=file
-id=16e6b160768b8c399314b7829cfc4d6037ec1339480
