+++ src/main/java/GradeApi/CourseController.java
@@ -26,131 +26,6 @@ public class CourseController {
     @Autowired
     JdbcTemplate jdbcTemplate;
 
-    // Gets the gen ed element 1 courses that student has taken.
-    public String getGenEdE1Remaining() {
-        List<Double> genEdE1Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E1"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE1Complete.size(); i++){
-            sum += genEdE1Complete.get(i);
-        }
-        sum = 9 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 1 satisfied.";
-        } else {
-            return "Gen Ed 1 needs " + sum + " credit hours.";
-        }
-
-    }
-
-    // Gets the gen ed element 2 courses that student has taken.
-    public String getGenEdE2Remaining() {
-        List<Double> genEdE2Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E2"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE2Complete.size(); i++){
-            sum += genEdE2Complete.get(i);
-        }
-        sum = 3 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 2 satisfied.";
-        } else {
-            return "Gen Ed 2 needs " + sum + " credit hours.";
-        }
-
-    }
-
-    // Gets the gen ed element 3 courses that student has taken.
-    public String getGenEdE3Remaining() {
-        List<Double> genEdE3Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E3"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE3Complete.size(); i++){
-            sum += genEdE3Complete.get(i);
-        }
-        sum = 6 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 3 satisfied.";
-        } else {
-            return "Gen Ed 3 needs " + sum + " credit hours.";
-        }
-
-    }
-
-    // Gets the gen ed element 4 courses that student has taken.
-    public String getGenEdE4Remaining() {
-        List<Double> genEdE4Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E4"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE4Complete.size(); i++){
-            sum += genEdE4Complete.get(i);
-        }
-        sum = 6 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 4 satisfied.";
-        } else {
-            return "Gen Ed 4 needs " + sum + " credit hours.";
-        }
-
-    }
-
-    // Gets the gen ed element 5 courses that student has taken.
-    public String getGenEdE5Remaining() {
-        List<Double> genEdE5Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E5"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE5Complete.size(); i++){
-            sum += genEdE5Complete.get(i);
-        }
-        sum = 6 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 5 satisfied.";
-        } else {
-            return "Gen Ed 5 needs " + sum + " credit hours.";
-        }
-
-    }
-
-    // Gets the gen ed element 5 courses that student has taken.
-    public String getGenEdE6Remaining() {
-        List<Double> genEdE6Complete = jdbcTemplate.query(
-                "SELECT credits FROM course WHERE requirement_satisfaction = ?", new Object[]{"Gen Ed E6"},
-                (rs, rowNum) -> rs.getDouble(1));
-
-        double sum = 0;
-        for (int i = 0; i < genEdE6Complete.size(); i++){
-            sum += genEdE6Complete.get(i);
-        }
-        sum = 6 - sum;
-        if ( sum <= 0 ){
-            return "Gen Ed 6 satisfied.";
-        } else {
-            return "Gen Ed 6 needs " + sum + " credit hours.";
-        }
-
-    }
-
-
-
-    /*
-    assume all concentrations will be:
-        General
-        Statistical Computing
-        Digital Forensics and Cybersecurity
-        Computer Technology
-        Interactive Multimedia
-        Artificial Intelligence in data Science
-     */
     @GetMapping("/getRemainingRequirements")
     public Object[] getRemainingRequirements(@RequestParam(value = "concentrationSelection", defaultValue = "General") String concentrationSelection) throws Exception {
         List<Assignment> assignments = jdbcTemplate.query(
@@ -159,21 +34,6 @@ public class CourseController {
         );
         return assignments.toArray();
     }
-
-
-    /*
-    Matthew:
-        Remember that course titles are 6 or 7 digits in length.
-        Assume that the digits are stored to the left, i.e.: "CSC190 ", instead of " CSC190" or "CSC 190".
-        Other requirements to check for:
-            Writing intensive course (the only course with a 7-digit title - always ends with a 'W')
-            Upper division coursework (courses ending with digits 300+)
-            ACCT requirement
-            Core
-            Supporting
-            concentration requirements
-            *120-hour requirement? Free electives? (we may not need to consider these - show me what you've got once you finish functions that check progress on other requirements)
-     */
 
     // get all courses
     @GetMapping("")
